🏗️ Refactor: Post Detail 페이지에 FSD 아키텍처 적용
