+++ clogging/src/app/(auth)/posts/page.tsx
@@ -1,0 +1,9 @@
+// 게시물 관련 페이지
+
+import React from 'react';
+
+const page = () => {
+  return <div></div>;
+};
+
+export default page;
