+++ clogging/src/app/api/posts/[id]/route.ts
@@ -1,0 +1,8 @@
+// 상세 게시물 API
+
+export async function GET(
+  req: Request,
+  { params }: { params: { id: string } },
+) {
+  // 게시물 상세 조회
+}
