+++ clogging/src/features/components/Comment/model/comment.ts
@@ -1,7 +1,7 @@
 'use client';
 
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { commentApi } from '@/entities/comment/api/commentApi';
+import { commentApi } from '../api/commentApi';
 
 export const useComments = (postId: string) => {
   return useQuery({
