+++ clogging/src/features/components/Post/model/post.ts
