+++ clogging/src/features/types.ts
@@ -1,3 +1,5 @@
+// 공통 타입 정의
+
 export interface Post {
   id: string;
   categoryId: number;
